splitup strds_management
